+++ Cargo.toml
@@ -15,3 +15,4 @@ sha2 = "0.10"
 argon2 = "0.5"
 
 data_access = {path = "./data_access"}
+data_models = {path = "./data_models"}
