+++ data_access/Cargo.toml
@@ -7,3 +7,5 @@ edition = "2021"
 sqlx = { version = "0.8.1", features = ["runtime-async-std-native-tls", "postgres", "chrono"] }
 serde = "1.0.208"
 chrono = { version = "0.4", features = ["serde"] }
+
+data_models = {path = "../data_models"}
