+++ data_models/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "data_models"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = "1.0.208"
