+++ data_models/src/auth.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct CreateAuthUserRequest {
+    pub username: String,
+    pub password: String,
+    pub email: String
+}
