+++ data_models/src/channel.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct CreateChannelRequest {
+    pub name: String,
+    pub guild_id: Option<i32>
+}
