+++ data_models/src/guilds.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct CreateGuildRequest {
+    pub name: String,
+    pub description: String,
+}
