+++ data_models/src/members.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct CreateMemberRequest {
+    pub user_id: i32,
+    pub guild_id: i32,
+}
