+++ data_models/src/messages.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct CreateMessageRequest {
+    pub channel_id: i32,
+    pub content: String,
+    pub attachments: Option<String>
+}
